Adding vgini to SConstruct
The sconstruct will now build the two vgini files,
move the executables to share and create a wrapper file.

This commit likely introduces new issues. I will test in the morning
as right now the tun driver is not loading because of a kernel upgrade.

TODO: move the .h files to the include directory.